# Improve the performance of snapshot_vec::extend